+++ src/app/modules/replay/replay-feature-settings.ts
@@ -1,6 +1,9 @@
 import { FeatureSettings } from '@app/feature';
 
 export interface ReplayFeatureSettings extends FeatureSettings {
+    replayCaptureManually?: boolean;
+    replayCaptureManuallyPastDuration?: number;
+    replayCaptureManuallyFutureDuration?: number;
     replayCaptureDeath?: boolean;
     replayCaptureDeathPastDuration?: number;
     replayCaptureDeathFutureDuration?: number;